update mock impl and move the segment event enum
