+++ packages/insomnia-app/app/common/__mocks__/analytics.ts
@@ -1,4 +1,5 @@
 // WARNING: changing this to `export default` will break the mock and be incredibly hard to debug. Ask me how I know.
-const trackEvent = jest.fn();
-const trackSegmentEvent = jest.fn();
-module.exports = { trackEvent, trackSegmentEvent };
+const _analytics = jest.requireActual('../analytics');
+_analytics.trackEvent = jest.fn();
+_analytics.trackSegmentEvent = jest.fn();
+module.exports = _analytics;
