+++ packages/insomnia-app/app/common/analytics.ts
@@ -19,7 +19,6 @@ import {
   isDevelopment,
 } from './constants';
 import { getScreenResolution, getUserLanguage, getViewportSize } from './electron-helpers';
-import { SegmentEvent } from './segment-event';
 
 const DIMENSION_PLATFORM = 1;
 const DIMENSION_VERSION = 2;
@@ -114,6 +113,23 @@ export async function getDeviceId() {
 }
 
 let segmentClient: Analytics | null = null;
+
+export enum SegmentEvent {
+  collectionCreate = 'Collection Created',
+  documentCreate = 'Document Created',
+  pluginExportLoadAllWokspace = 'Plugin export loading all workspace',
+  pluginExportLoadWorkspacesInSpace = 'Plugin export loading workspaces for active space',
+  requestCreate = 'Request Created',
+  requestExecute = 'Request Executed',
+  spaceLocalCreate = 'Local Space Created',
+  spaceLocalDelete = 'Local Space Deleted',
+  testSuiteCreate = 'Test Suite Created',
+  testSuiteDelete = 'Test Suite Deleted',
+  unitTestCreate = 'Unit Test Created',
+  unitTestDelete = 'Unit Test Deleted',
+  unitTestRun = 'Ran Individual Unit Test',
+  unitTestRunAll = 'Ran All Unit Tests',
+}
 
 export async function trackSegmentEvent(event: SegmentEvent, properties?: Record<string, any>) {
   const settings = await models.settings.getOrCreate();
