--- packages/insomnia-app/app/common/segment-event.ts
@@ -1,16 +1,0 @@
-export enum SegmentEvent {
-  collectionCreate = 'Collection Created',
-  documentCreate = 'Document Created',
-  pluginExportLoadAllWokspace = 'Plugin export loading all workspace',
-  pluginExportLoadWorkspacesInSpace = 'Plugin export loading workspaces for active space',
-  requestCreate = 'Request Created',
-  requestExecute = 'Request Executed',
-  spaceLocalCreate = 'Local Space Created',
-  spaceLocalDelete = 'Local Space Deleted',
-  testSuiteCreate = 'Test Suite Created',
-  testSuiteDelete = 'Test Suite Deleted',
-  unitTestCreate = 'Unit Test Created',
-  unitTestDelete = 'Unit Test Deleted',
-  unitTestRun = 'Ran Individual Unit Test',
-  unitTestRunAll = 'Ran All Unit Tests',
-}
