+++ packages/insomnia-app/app/ui/redux/modules/__tests__/git.test.tsx
@@ -8,9 +8,8 @@ import { mocked } from 'ts-jest/utils';
 
 import { globalBeforeEach } from '../../../../__jest__/before-each';
 import { reduxStateForTest } from '../../../../__jest__/redux-state-for-test';
-import {  trackEvent, trackSegmentEvent } from '../../../../common/analytics';
+import { SegmentEvent, trackEvent, trackSegmentEvent } from '../../../../common/analytics';
 import { ACTIVITY_SPEC } from '../../../../common/constants';
-import { SegmentEvent } from '../../../../common/segment-event';
 import * as models from '../../../../models';
 import { gitRepositorySchema } from '../../../../models/__schemas__/model-schemas';
 import { BASE_SPACE_ID } from '../../../../models/space';
