+++ packages/insomnia-app/app/ui/redux/modules/__tests__/space.test.js
@@ -3,9 +3,8 @@ import thunk from 'redux-thunk';
 
 import { globalBeforeEach } from '../../../../__jest__/before-each';
 import { reduxStateForTest } from '../../../../__jest__/redux-state-for-test';
-import { trackEvent, trackSegmentEvent } from '../../../../common/analytics';
+import { SegmentEvent, trackEvent, trackSegmentEvent } from '../../../../common/analytics';
 import { ACTIVITY_HOME } from '../../../../common/constants';
-import { SegmentEvent } from '../../../../common/segment-event';
 import * as models from '../../../../models';
 import { BASE_SPACE_ID } from '../../../../models/space';
 import { getAndClearShowAlertMockArgs, getAndClearShowPromptMockArgs } from '../../../../test-utils';
